- changed: If the IWAD contains a MENUDEF lump the one in zdoom.pk3 should not print texture warnings.
The reason is that in such a case it is very likely that the IWAD defines its own menu and will most likely not provide all assets for the base definitions. See 'Adventures of Square' for an example.

(cherry picked from commit 6c37ab231099f0adf2884b8ad7955aa8e154c31e)

# Conflicts:
#	src/menu/menudef.cpp