Revert "- Addressed crash on exit with MinGW. This is a hack and i've reported it as a bug in the MinGW C runtime."
This reverts commit 558907272907194331d03a3abf394080ae46c92a.

There are two different crashing bugs in MinGW, a crash related to destruction of generic objects and another one with thread_local objects both with race conditions. For the hack to be fully effective you'd need to remove the thread_local keyword as well. This was just a workaround for testing.