- changed the stencil cap drawer to only cover the area which is actually used by the portal.
This will now both exclude floor caps when only ceiling elements are used and everything outside the bounding box of active portal lines.
Hopefully this is enough to fix the issues with portal caps but of course it is not foolproof if someone just makes the right setup.

# Conflicts:
#	src/hwrenderer/scene/hw_portal.cpp
#	src/hwrenderer/scene/hw_portal.h
#	src/hwrenderer/scene/hw_walls.cpp
#	src/r_data/r_sections.cpp
#	src/r_data/r_sections.h

- Using FBoundingBox instead of BoundingRect for the implementation.
- Added Contains methods to FBoundingBox.
Doesn't compile yet. (drfrag)